おそらくshell変更前の最後のコミット ブラウンリセットをOFF pi(一応dも)のパラメータ調整、get,setの増加、その他関数を調整。 このバージョンで使うときには(不要な)デバック用関数等を消すほうがいい

+++ MDU_COMON.X/Driver/pwm.h
@@ -32,7 +32,7 @@ inline void pwm_shutdown(bool fag);
 int16_t timer_flag();
 void pwm_pid_period(uint16_t);
 int16_t pid_rate();
-void get_mc_period(uint32_t *,uint16_t *);
+void get_mc_period(uint32_t *,uint32_t *);
 
 void set_event(event_t);
 #endif	/* PWM_H */
